Feat: 온보딩 - 닉네임 step 과 리스트 생성 step 분리
